requirements.txt
#Python version 3.13
mysql-connector-python